This commit introduces two things :- [1] Fixes chunk processing for process_failed_items, keeping the order of items the same, as they were before the invokation.
[2] Moves the responsibility of re-queueing items in the
failed queue to process_failed_items itself, by default. If
the callback returns 0 or dies, while processing a failed
item, it is requeued in the failed queue. We should probably
limit this by introducing a max_retries on the item level
so as not to keep processing the same failed items repeatedly.